+++ NEWS.md
@@ -1,6 +1,6 @@
 # rasterpic (development version)
 
--   Add support for `SpatVector`, `SpatExtent`.
+-   Add support for `SpatVector`, `SpatExtent`, `sf::st_bbox()`.
 
 # rasterpic 0.1.0